pyproject.toml
[project]
name = "leadflow-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "aiogram (>=3.26.0,<4.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "google-auth (>=2.49.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]

[tool.black]
line-length = 88
force-exclude = '''
/(
    \.git

  | \.venv
  | migrations/versions  # Игнорировать папку с версиями миграций
)/
'''

[tool.isort]
profile = "black"
skip_glob = ["migrations/versions/*"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
